refactor: Moved Project card rendering into a separate component
